Set JT_ENV to use truffleruby-jvm-ce by default
A recent change to TruffleRuby [0] means that the `JT_ENV` environment
variable will be respected by both `jt build` and `jt ruby` as the
default value of their `--env` and `--use` flags respectively. `jvm-ce`
is what we normally want to use in development so it’s convenient for it
to be our Spin default.

I’ve used `export` [1,2] to set this environment variable so that its
value is inherited by any child processes of the login shell (e.g. shell
scripts, subshells) because that’s probably the behaviour that we want.

[0] https://github.com/oracle/truffleruby/commit/3e6908c5d06e9d4ef24380e429db0d9482cbd0a7
[1] https://www.gnu.org/software/bash/manual/bash.html#index-export
[2] https://tldp.org/LDP/abs/html/internal.html#EXPORTREF